moving play page logic to a separate file
